Fix: Pass project name to "docker compose"
Previously dkp did not export the volumes of the project if the
project's directory was not the same as the project's name. This was
because "docker compose config" assumed that the project's name is the
directory, and created the volume names based on that.

For example, if I have a directory called "my-dir", but I always call
"docker compose" with "-p my-project" (or I set
COMPOSE_PROJECT_NAME=my-project), and I have volume called "my-volume",
then the volume's real name will be "my-project_my-volume". However,
when I issued "dkp my-project" (without setting COMPOSE_PROJECT_NAME),
it didn't pass "-p my-project" to docker compose, so docker compose
printed a config that used "my-dir_my-volume" as the volume name, and
dkp tried to export that volume.

After this commit, we pass the project name to docker compose, so it
will print the volumes with the correct names, so dkp can export them.