Export non-prebuilt images
Let's say that we use the following compose file for the test-data
project:

    services:
      dummy:
        image: busybox
      web:
        build:
          context: .
          dockerfile_inline: |
            FROM nginx:alpine

When the user starts the project, Compose builds and stores the nginx
image as "test-data_web:latest".

Before this commit, dkp exported only the busybox image.

After this commit, the default behavior stays the same, but if the user
passes the "--all-images" option to dkp, it will also export the
"test-data_web:latest" image.