Multiple env files (and other improvements) (#2)
* fix typos

* Use precise match when finding the project name

Previously, the "docker compose ls ... Name={project_name}" filter found
projects whose name contained {project_name} as a substring. For
example, if the project name was "my_project", it also found
"this_is_my_project".

The Docker documentation mentions that "You can also filter for a
substring in a name"
(https://docs.docker.com/reference/cli/docker/container/ls/), but for
some reason it's not documented that the specified "name" is actually
used as a regular expression
(https://github.com/docker/docs/issues/11767).

This commit modifies dkp to find (and archive) only the specified
project.

According to
https://docs.docker.com/compose/project-name/#set-a-project-name,
"project names must contain only lowercase letters, decimal digits,
dashes, and underscores", so we can safely use the project name as a
regular expression.

* Fix copying multiple env files

Previously the env file copying worked properly only when there was one
env file and it was called ".env".

*   If there were multiple env files, including .env, dkp copied the
    content of .env into all of them (because "info.env_file" always
    pointed to ".env").

*   If there were multiple env files, but no ".env", dkp crashed:

    FileNotFoundError: [Errno 2] No such file or directory:
    '.../test-data/.env'

Now dkp copies every .env and *.env file with its own content.

* Fix: Pass project name to "docker compose"

Previously dkp did not export the volumes of the project if the
project's directory was not the same as the project's name. This was
because "docker compose config" assumed that the project's name is the
directory, and created the volume names based on that.

For example, if I have a directory called "my-dir", but I always call
"docker compose" with "-p my-project" (or I set
COMPOSE_PROJECT_NAME=my-project), and I have volume called "my-volume",
then the volume's real name will be "my-project_my-volume". However,
when I issued "dkp my-project" (without setting COMPOSE_PROJECT_NAME),
it didn't pass "-p my-project" to docker compose, so docker compose
printed a config that used "my-dir_my-volume" as the volume name, and
dkp tried to export that volume.

After this commit, we pass the project name to docker compose, so it
will print the volumes with the correct names, so dkp can export them.

* Support multiple env files

Previously dkp assumed that:

*   The .env and *.env files next to the first compose file were the env
    files shall be copied into the archive.

*   The "docker compose" commands don't require "--env-file <file>"
    options to work.

After this commit, the users can specify a list of env files with the
option "--env-file <file1> <file2>...".

*   These (and only these) env files are copied into the archive.

*   These env files are passed to the "docker compose" commands.

If no "--env-file" option is specified, dkp works the same way as
before.

* Export non-prebuilt images

Let's say that we use the following compose file for the test-data
project:

    services:
      dummy:
        image: busybox
      web:
        build:
          context: .
          dockerfile_inline: |
            FROM nginx:alpine

When the user starts the project, Compose builds and stores the nginx
image as "test-data_web:latest".

Before this commit, dkp exported only the busybox image.

After this commit, the default behavior stays the same, but if the user
passes the "--all-images" option to dkp, it will also export the
"test-data_web:latest" image.

* Fix typing problems

Before:

    $ mypy dkp.py
    dkp.py:226: error: Need type annotation for "args" (hint: "args: List[<type>] = ...")  [var-annotated]
    dkp.py:359: error: Incompatible types in assignment (expression has type "Path", variable has type "str")  [assignment]
    dkp.py:364: error: Unsupported left operand type for / ("str")  [operator]
    dkp.py:372: error: Unsupported left operand type for / ("str")  [operator]
    dkp.py:377: error: Unsupported left operand type for / ("str")  [operator]
    dkp.py:423: error: Argument 1 to "gen_scripts" has incompatible type "str"; expected "Path"  [arg-type]
    dkp.py:431: error: Argument 3 to "encrypt" has incompatible type "Optional[str]"; expected "str"  [arg-type]
    Found 7 errors in 1 file (checked 1 source file)

After:

    $ mypy dkp.py
    Success: no issues found in 1 source file

* Improve copying env files

1.  Previously when the user passed the `--env-file` option to dkp, and
    the compose file used `*.env` files with the `env_file` attribute,
    dkp did not copy these `*.env` files.

    Now dkp copies the `*.env` files in this case.

    This is still not perfect because dkp's heuristic is that it copies
    all `*.env` files that it finds in the project root directory.
    Sometimes a file is copied unnecessarily (when it is not used), and
    sometimes a file is not copied that should be (when an env file does
    not have a *.env filename, or it is located in a subdirectory).

    A more comprehensive solution would be to collect the env file names
    from the `env_file` attributes of the composed files.

2.  Previously dkp copied all env files into the target root directory
    (`work_dir`). This was fine for those env files are are in the
    project root directory, but this was not fine for other env files.

    Now:

    *   dkp skips the env files that are outside the project directory.
        dkp could not restore this files on the target machine, so the
        user should take care about them. That's why now dkp prints
        messages about skipping these files.

    *   dkp re-creates the directory structure for those env files that
        are in subdirectories inside the project directory.

* Make passphrase optional

Previously specifying a passphrase for the pack was mandatory. This
could be done either by using the "--passphrase"/"-p" option, or by
using the PASSPHRASE environment variable.

This commit makes the passphrase optional. This way the users can create
unencrypted packages, and they don't need GPG to be installed.